requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-decouple==3.8
django-crispy-forms==2.0
crispy-bootstrap5==0.7
Pillow==10.1.0
django-extensions==3.2.3
reportlab==4.0.7
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
cloudinary==1.36.0
django-cloudinary-storage==0.3.0